requirements.txt
boto3~=1.14.60
elasticsearch~=7.7.0
pandas~=1.0.5
psycopg2-binary~=2.8.5
pyarrow~=0.17.0
pytest~=5.4.3
PyYAML~=5.3.1
redis~=3.5.2
requests-aws4auth~=0.9
retry~=0.9.2
s3fs~=0.4.2
SQLAlchemy~=1.3.17
sentry-sdk~=0.7.9